# api/openapi.yaml
openapi: 3.0.3
info:
  title: Harmattan API
  version: 0.1.0
paths:
  /sign-up:
    post:
      tags: [authentication]
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
              required: [email, password]
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_token:
                    type: string
                required: [id_token]
        default:
          $ref: "#/components/responses/error"
  /sign-in:
    post:
      tags: [authentication]
      operationId: signIn
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
              required: [email, password]
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_token:
                    type: string
                required: [id_token]
        default:
          $ref: "#/components/responses/error"
  /projects:
    post:
      tags: [projects]
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 20
                color:
                  type: string
                  minLength: 7
                  maxLength: 7
              required: [name, color]
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/project"
        default:
          $ref: "#/components/responses/error"
      security:
        - bearerAuth: []
    get:
      tags: [projects]
      operationId: listProjects
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/projects"
        default:
          $ref: "#/components/responses/error"
      security:
        - bearerAuth: []
  /projects/{projectID}:
    parameters:
      - $ref: "#/components/parameters/projectID"
    patch:
      tags: [projects]
      operationId: updateProject
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 20
                color:
                  type: string
                  minLength: 7
                  maxLength: 7
                is_archived:
                  type: boolean
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/project"
        default:
          $ref: "#/components/responses/error"
      security:
        - bearerAuth: []
  /health:
    get:
      tags: [monitoring]
      operationId: checkHealth
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  revision:
                    type: string
                required: [revision]
        default:
          $ref: "#/components/responses/error"
components:
  schemas:
    project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        is_archived:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, color, is_archived, created_at, updated_at]
    projects:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/project"
        has_next:
          type: boolean
      required: [projects, has_next]
  parameters:
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    projectID:
      name: projectID
      in: path
      required: true
      schema:
        type: string
        minLength: 26
        maxLength: 26
  responses:
    error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
              message:
                type: string
            required: [code, message]
  securitySchemes:
    bearerAuth:
      type: http
      scheme: Bearer
tags:
  - name: authentication
  - name: projects
  - name: monitoring